fixerrors

+++ srcs/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: hfilipe- <hfilipe-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 15:07:02 by hfilipe-          #+#    #+#             */
-/*   Updated: 2024/10/21 21:43:06 by hfilipe-         ###   ########.fr       */
+/*   Updated: 2024/10/23 21:24:26 by hfilipe-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
 	size_t	srclen;
